refactor: move mock factories to mocks/ directories
Move 7 mock factory files to mocks/ subdirectories to comply with
the mock factory placement conventions in basic-rule.md:
- createMockTabsService.ts
- createMockCurrentUrlService.ts
- createMockPopupService.ts
- createMockSelectedPageTextRepository.ts
- createMockDomRootChecker.ts
- createMockElementFactory.ts
- createMockTabsGateway.ts

Update import paths in 26 test files that reference these mocks.
Remove the "未対応モック一覧" section from basic-rule.md.
