+++ docs/coding-standards/tests/unit/common-rule/basic-rule.md
@@ -169,14 +169,6 @@ const rule = RewriteRule.fromParams(1, {
 
 **注意**: 新規テスト作成時は必ずファクトリーメソッドを使用すること。
 
-## mocks/ ディレクトリ未配置モック
-
-規約に準拠していないモックファクトリは [user-story-004](../../../user-stories/user-story-004/README.md) で対応予定。
-
-**注意**: 新規モック作成時は必ず `mocks/` ディレクトリに配置すること。
-
----
-
 ## Reactテストユーティリティのインポート
 
 ### 規約
