+++ host-frontend-root/frontend-src-root/tests/unit/application-business-rules/interactors/DeleteRuleInteractor/execute/partial-success-cases.test.ts
@@ -2,7 +2,7 @@
  * DeleteRuleInteractor.execute - 部分的成功テスト
  * 1. tabsGateway.reloadMatchingTabsでエラーが発生した場合、presentが先に呼び出された後presentErrorが呼び出される
  */
-import { createMockTabsGateway } from 'tests/frameworks-and-drivers/browser/ChromeTabsGateway/createMockTabsGateway';
+import { createMockTabsGateway } from 'tests/frameworks-and-drivers/browser/ChromeTabsGateway/mocks/createMockTabsGateway';
 import { createMockRewriteRuleRepository as createMockRepository } from 'tests/unit/application/ports/IRewriteRuleRepository/mocks/createMockRewriteRuleRepository';
 import { createMockPresenter } from 'tests/unit/application-business-rules/interactors/DeleteRuleInteractor/mocks/createMockPresenter';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
