+++ host-frontend-root/frontend-src-root/tests/unit/domain/entities/DomDiffer/normal-replacement.test.ts
@@ -1,4 +1,4 @@
-import { createMockElementFactory } from 'tests/unit/domain/ports/IElementFactory/createMockElementFactory';
+import { createMockElementFactory } from 'tests/unit/domain/ports/IElementFactory/mocks/createMockElementFactory';
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 
 import { DomDiffer } from 'src/domain/entities/DomDiffer';
