+++ host-frontend-root/frontend-src-root/tests/unit/domain/entities/ParserContextStrategy/ParserContextStrategyFactory/createContainer/normal-cases.test.ts
@@ -1,4 +1,4 @@
-import { createMockElementFactory } from 'tests/unit/domain/ports/IElementFactory/createMockElementFactory';
+import { createMockElementFactory } from 'tests/unit/domain/ports/IElementFactory/mocks/createMockElementFactory';
 import { describe, expect, it } from 'vitest';
 
 import { ParserContextStrategyFactory } from 'src/domain/entities/ParserContextStrategy';
