+++ host-frontend-root/frontend-src-root/tests/unit/domain/ports/IDomRootChecker/mocks/createMockDomRootChecker.ts
